fix(runtime): preserve snapshot failure responses and test isolation
